refactor : Car 객체 구현
